+++ src/frontend/src/modules/builder/BuilderPriceCounter.vue
@@ -1,0 +1,14 @@
+<template>
+  <div class="content__result">
+    <p>Итого: 0 ₽</p>
+    <button type="button" class="button" disabled>Готовьте!</button>
+  </div>
+</template>
+
+<script>
+export default {
+  name: "BuilderPriceCounter",
+};
+</script>
+
+<style scoped></style>
